requirements.txt
pandas>=1.5.0
numpy>=1.22.0
requests>=2.28.0
pyarrow>=8.0.0
python-dateutil>=2.8.2
matplotlib>=3.6.0  # For backtest visualization
scipy>=1.9.0       # For technical indicators
python-dotenv>=1.0.0 # For environment variables
